--- backend/src/config/db.js
@@ -1,13 +1,0 @@
-const mongoose = require("mongoose");
-
-const connectDB = async () => {
-  try {
-    await mongoose.connect("mongodb://127.0.0.1:27017/movielist");
-    console.log("✅ Connected to MongoDB");
-  } catch (error) {
-    console.error("MongoDB connection error:", error.message);
-    process.exit(1);
-  }
-};
-
-module.exports = connectDB;
