--- backend/src/routes/auth.js
@@ -1,40 +1,0 @@
-const express = require("express");
-const bcrypt = require("bcryptjs");
-const User = require("../models/User");
-
-const router = express.Router();
-const saltRounds = 10;
-
-// Register
-router.post("/users", async (req, res) => {
-  const { username, password } = req.body;
-  try {
-    const existing = await User.findOne({ username });
-    if (existing) return res.status(400).json({ error: "User already exists" });
-
-    const hashedPassword = await bcrypt.hash(password, saltRounds);
-    const user = new User({ username, password: hashedPassword });
-    await user.save();
-    res.json({ message: "User registered", user });
-  } catch (err) {
-    res.status(500).json({ error: "Server error" });
-  }
-});
-
-// Login
-router.post("/login", async (req, res) => {
-  const { username, password } = req.body;
-  try {
-    const user = await User.findOne({ username });
-    if (!user) return res.status(400).json({ error: "Invalid credentials" });
-
-    const isMatch = await bcrypt.compare(password, user.password);
-    if (!isMatch) return res.status(400).json({ error: "Invalid credentials" });
-
-    res.json({ message: "Login successful", user });
-  } catch (err) {
-    res.status(500).json({ error: "Server error" });
-  }
-});
-
-module.exports = router;
