--- backend/src/routes/user.js
@@ -1,16 +1,0 @@
-const express = require("express");
-const User = require("../models/User");
-const router = express.Router();
-
-// Get User
-router.get("/user/:username", async (req, res) => {
-  try {
-    const user = await User.findOne({ username: req.params.username });
-    if (!user) return res.status(404).json({ error: "User not found" });
-    res.json(user);
-  } catch (err) {
-    res.status(500).json({ error: "Failed to fetch user" });
-  }
-});
-
-module.exports = router;
