+++ package.json
@@ -10,14 +10,8 @@
   },
   "dependencies": {
     "axios": "^1.8.3",
-    "bcryptjs": "^3.0.2",
-    "body-parser": "^2.2.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.5.0",
     "elecmovie-tailwind-starter": "file:",
-    "express": "^5.1.0",
     "lucide-react": "^0.483.0",
-    "mongoose": "^8.15.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
@@ -52,5 +46,5 @@
   "bugs": {
     "url": "https://github.com/rodvincent1/FINALMOVIELIST-ELECTIVE3/issues"
   },
-  "homepage": "https://github.com/rodvincent1/FINALMOVIELIST-ELECTIVE3#readme"
+  "homepage": ""
 }
